refactor: remove shared barrel imports across tool-ui
