+++ components/tool-ui/approval-card/approval-card.tsx
@@ -2,9 +2,9 @@
 
 import * as React from "react";
 import { cn, Separator } from "./_adapter";
-import type { ApprovalCardProps, ApprovalDecision } from "./schema";
-import { ActionButtons } from "../shared";
-import type { Action } from "../shared";
+import type { ApprovalCardProps, ApprovalDecision } from "./schema";import { ActionButtons } from "../shared/action-buttons";
+import { type Action } from "../shared/schema";
+
 import { icons, Check, X } from "lucide-react";
 
 type LucideIcon = React.ComponentType<{ className?: string }>;
