+++ components/tool-ui/code-block/error-boundary.tsx
@@ -1,5 +1,3 @@
-"use client";
-
-import { createToolUiErrorBoundary } from "../shared";
+"use client";import { createToolUiErrorBoundary } from "../shared/error-boundary";
 
 export const CodeBlockErrorBoundary = createToolUiErrorBoundary("CodeBlock");
