+++ components/tool-ui/code-block/schema.ts
@@ -1,12 +1,5 @@
-import { z } from "zod";
-import {
-  ToolUIIdSchema,
-  ToolUIReceiptSchema,
-  ToolUIRoleSchema,
-  SerializableActionSchema,
-  SerializableActionsConfigSchema,
-  defineToolUiContract,
-} from "../shared";
+import { z } from "zod";import { defineToolUiContract } from "../shared/contract";
+import { SerializableActionSchema, SerializableActionsConfigSchema, ToolUIIdSchema, ToolUIReceiptSchema, ToolUIRoleSchema } from "../shared/schema";
 
 export const CodeBlockPropsSchema = z.object({
   id: ToolUIIdSchema,
