+++ components/tool-ui/image-gallery/error-boundary.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { createToolUiErrorBoundary } from "../shared";
+"use client";import { createToolUiErrorBoundary } from "../shared/error-boundary";
 
 export const ImageGalleryErrorBoundary =
   createToolUiErrorBoundary("ImageGallery");
