+++ components/tool-ui/image/image.tsx
@@ -2,12 +2,9 @@
 "use client";
 
 import * as React from "react";
-import { cn } from "./_adapter";
-import {
-  ActionButtons,
-  normalizeActionsConfig,
-  type ActionsProp,
-} from "../shared";
+import { cn } from "./_adapter";import { ActionButtons } from "../shared/action-buttons";
+import { normalizeActionsConfig, type ActionsProp } from "../shared/actions-config";
+
 import {
   RATIO_CLASS_MAP,
   getFitClass,
