+++ components/tool-ui/parameter-slider/error-boundary.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { createToolUiErrorBoundary } from "../shared";
+"use client";import { createToolUiErrorBoundary } from "../shared/error-boundary";
 
 export const ParameterSliderErrorBoundary =
   createToolUiErrorBoundary("ParameterSlider");
