+++ components/tool-ui/question-flow/schema.ts
@@ -1,10 +1,6 @@
 import { z } from "zod";
-import type { ReactNode } from "react";
-import {
-  ToolUIIdSchema,
-  ToolUIRoleSchema,
-  defineToolUiContract,
-} from "../shared";
+import type { ReactNode } from "react";import { defineToolUiContract } from "../shared/contract";
+import { ToolUIIdSchema, ToolUIRoleSchema } from "../shared/schema";
 
 export const QuestionFlowOptionSchema = z.object({
   id: z.string().min(1),
