+++ components/tool-ui/video/error-boundary.tsx
@@ -1,5 +1,3 @@
-"use client";
-
-import { createToolUiErrorBoundary } from "../shared";
+"use client";import { createToolUiErrorBoundary } from "../shared/error-boundary";
 
 export const VideoErrorBoundary = createToolUiErrorBoundary("Video");
