+++ components/tool-ui/x-post/schema.ts
@@ -1,5 +1,4 @@
-import { z } from "zod";
-import { defineToolUiContract } from "../shared";
+import { z } from "zod";import { defineToolUiContract } from "../shared/contract";
 
 export const XPostAuthorSchema = z.object({
   name: z.string(),
