+++ package.json
@@ -12,7 +12,7 @@
     "registry:build": "tsx scripts/build-tool-ui-registry.ts",
     "registry:check": "pnpm registry:build && git diff --exit-code -- public/r",
     "lint": "eslint .",
-    "lint:ci": "eslint components/tool-ui/data-table components/tool-ui/message-draft components/tool-ui/option-list components/tool-ui/plan components/tool-ui/progress-tracker components/tool-ui/weather-widget/effects/glass-panel-svg.tsx lib/registry/tool-ui-registry.ts lib/tests/registry lib/tests/tool-ui lib/tests/setup scripts/build-tool-ui-registry.ts eslint.config.ts vitest.config.ts",
+    "lint:ci": "eslint components/tool-ui lib/registry/tool-ui-registry.ts lib/tests/registry lib/tests/tool-ui lib/tests/setup scripts/build-tool-ui-registry.ts eslint.config.ts vitest.config.ts",
     "lint:fix": "eslint . --fix",
     "typecheck": "tsc --noEmit",
     "test": "vitest run",
